+++ cmd/gon/main.go
@@ -45,7 +45,7 @@ func run(args []string) int {
 		usage()
 		return 0
 	case "version", "-version", "--version":
-		fmt.Println("gon version 1.3.0")
+		fmt.Println("gon version 1.4.0")
 		return 0
 	}